General improvements.
Isolated the variables of the DOM script by placing it correctly in another scope and setting the downloadURL as a property of the parented <script> tag instead of injecting it in a second DOM script.

The reliability of the page change detection has been improved and videos should now properly be played without the need of refreshing the page.

Also general formatting consistencies have been fixed.